Correct Idea Kotlin plugin dependency (#2715)
* Build and test with Kotlin version bundled with IntelliJ

Plugin dependencies can't be pinned to a specific version, so building &
testing on newer Kotlin versions can mask issues that will occur when
running the plugin on older IntelliJ versions which are using older Kotlin
plugin versions.

* Remove kotlin-reflect dependency

kotlin-reflect is bunded with IntelliJ, and using a mismatched dependency
can lead to errors.

* Compile project against Kotlin API version 1.3

This is to maintain compatibility with bundled Kotlin 1.3.70 in IntelliJ
2020.2.

* Remove unused plugin dependencies

These pull in the Groovy and properties plugins during compile, build &
test, but they are not used by the SQLDelight plugin and aren't declared
as required or optional dependencies in plugin.xml.

* Align Gradle plugin name to identifier used in plugin.xml

* Explain reason for targeting Kotlin API version 1.3